SWARM-1964: MP-JWT should support public keys from JWKS URI
Motivation
----------
It is common that token issuers provide a convenient JWKS URI to distribute public keys (for signature-verification). Wildfly Swarm's MP-JWT fraction should be able to take advantage of these JWKS URIs rather than hard-code a public key in the configuration. The major benefits are that several issuers are supported (through "kid"-matching) and that expiring/changing public keys do not require modifications to an application's configuration.

Modifications
-------------
Several classes in the MP-JWT fraction (configuration-related classes, signature-validation of JWT) and an Arquillian-based testsuite to verify different configuration combinations and key-provider-mechanisms.

Result
------
The MP-JWT fraction can no be configured through MP-Config and Fraction-config to support a JWKS address for public keys ('jwksUri' and 'jwksRefreshInterval' have been added). Several signer keys are now supported (through "kid"-matching) and it's no longer mandatory to provide a static public key just as long as a JWKS address is provided. If both JWKS URIs and a static key is present, the latter is used.

Signed-off-by: Daniel Pfeifer <daniel@pfeifer.io>