Simplify POM files
Motivation
----------
When testing out changes locally, I tend to use `mvn versions:set -DnewVersion=xxx` a lot. As a habit, I tend to look at the changed files. In the current form, there are a lot of changes that show up in two POM files in particular - the root POM and the Netflix POM.

Modifications
-------------
* Use ${project.version} instead of an explicit version detail.
* Set `wildfly-swarm` as the parent of `build-resources` & `maven-enforcer-pattern-size`.
* Moved the enforcer plugin configuration from root to one level below. This is needed to avoid circular dependencies during POM evaluation.

Result
------
* Verified that the build is going through properly.
* Generated a custom build and verified that the new behavior is working properly.