SWARM-SWARM-1976: Repackage war archives in gradle plugin
Motivation
----------
The maven plugin systematically repackages war files.
The current gradle plugin (2018.4.1) never repackages.
Therefore, in hollow mode, the generated war is not usable with the hollow jar.

Modifications
-------------
If a war task exists, scan for all ".war" files in the artifacts and repackage them

Result
------
The resulting war file(s) will be usable for the hollow jar.