SWARM-1971: make OpenApi work with arrays and generics
Motivation
----------
Currently the application cannot be deployed if someone exposes an
endpoint that accepts an array or a generic type.
This commit makes the OpenApi fraction usable in this case.

Modifications
-------------
JandexUtil in microprofile-openapi fixed not to use (unncessary)
`.asClassType`.
Also, a test was added to check if application behaves better.

Result
------
An application that has endpoints consuming arrays or generics can now
use the OpenApi fraction.

Please note that the changes do not fix the data the openapi endpoint returns for arrays.
They are only intended to make OpenApi fraction start with it.
That's why the test for arrays is ignored