SWARM-1993: support user classes as config sources for MP Config
Motivation
----------
Before the changes, the only possible value for `swarm.microprofile.config.config-sources.KEY.attribute-class` property was:
```
 swarm.microprofile.config.config-sources.KEY.attribute-class:
    module: some.module
    name: my.package.MyClass
```

This change bring the possibility to use classes by class name, e.g:
```
swarm:
   microprofile:
      config:
         config-sources:
            default:
               attribute-class:
                com.test.microprofile.config.TestConfigSource
```

Modifications
-------------
MicroProfileConfigDeploymentProcessor has been added.
The processor checks if the value defined `swarm.microprofile.config.config-sources.KEY.attribute-class`
is a `String` and if it is, assumes it is a name of a class and adds a microprofile config service definition pointing to it.

Result
------
The YAML configuration for attribute-class works as described in the Motivation section