Fixes #8 - conf/ dir support
If executed via the run.sh or run.bat, then the conf/ dir alongside
the bin/ dir is added as a config location to hold .properties or .yml
configuration.

Additionally, plugin now creates that directory with a README.txt to
make it obvious.

Also, renamed core to kernel for logging stuff.