THORN-2036: Added autodetect for MicroProfile Metrics, OpenApi and RestClient
Motivation
----------
All MicroProfile APIs should work the same way. Adding autodetect for the APIs that does not have it.

Modifications
-------------
Added a detect package and a class that extends PackageFractionDetector for Metrics, OpenApi and RestClient

Result
------
When building a fat or hollow jar, the build will include the Metrics, OpenApi and RestClient fractions if your code use it.

Changes to be committed:
	new file:   fractions/microprofile/microprofile-metrics/src/main/java/org/wildfly/swarm/microprofile/metrics/detect/MetricsPackageDetector.java
	new file:   fractions/microprofile/microprofile-openapi/src/main/java/org/wildfly/swarm/microprofile/openapi/detect/OpenApiPackageDetector.java
	new file:   fractions/microprofile/microprofile-restclient/src/main/java/org/wildfly/swarm/microprofile/restclient/detect/RestClientPackageDetector.java