Make consul TTL check timeout configurable
Motivation
----------
In topology-consul when a service is advertised the advertiser automatically creates a TTL check for every service. This check is created with a fixed timeout of 3s. In an endless loop the advertiser is then cycling over all services and sends the alive check to consul - but with a 2s sleep timer at the end of each round.
This results in only 1s time window for the alive check. If there are many services to process or slight network delay the alive ping will not arrive in time resulting in flapping reachability of the service in consul.

Modifications
-------------
Added a configurable Long value 'swarm.topology.consul.ttl' which will get passed via AgentActivator to Advertiser and then used as value for the ttl in RegCheck. The value is Defaultable<Long> and has a default value of 3.

Result
------
The deafult behaviour is not changed but the ttl for the consul check can now be configured to a higher value if needed.