THORN-2143: MP JWT: using * in @RolesAllowed doesn't work
Motivation
----------
After the recent refactoring of MP JWT authorization, annotating
a resource method with `@RolesAllowed("*")` no longer works.
This is inconsistent with previous behavior and also with
how security annotations behave by default.

Modifications
-------------
If one of the allowed roles is `*`, then access is allowed
for all logged-in users.

Result
------
Using `@RolesAllowed("*")` works as expected again.