THORN-2110 followup: improve the principal leak test
Motivation
----------
The `PrincipalLeakTest` assumed that the number of worker threads
is always 64. That, however, isn't true -- it's actually derived
from the number of currently available CPU cores.

It's much better to explicitly configure the IO subsystem
to make sure only one thread is ever used. With a single thread,
there are no questions about how threads are chosen etc.

Modifications
-------------
Configure the IO subsystem in this test to restrict the number
of worker threads to 1 and adjust the test accordingly.

Result
------
More robust test.