THORN-2148: Added unit test to reproduce issue with "hierarchical" properties and implemented a fix. Github issue: #1094
Motivation
----------
Fix the issue described in THORN-2148: if properties in the properties file have a hierarchy and a node has both children and a value of its own, the result is undefined.

Modifications
-------------
Created a unit test to reproduce the issue (ConfigViewFactoryTest.java) and another to isolate the issue as not being in the command line processing (CommandLineTest.java). Then
fixed the issue in org.wildfly.swarm.container.config.ConfigNode.

Result
------
Fix is fully backwards compatible and does not affect users not using a properties file or not using a properties file with a hierarchy.