THORN-2137: properly handle /metric requests without accept headers
Motivation
----------
During migration to SmallRye I introduced a bug that resulted in
throwing an NPE on metrics request without headers.
This change fixes the bug.

Modifications
-------------
acceptHeaders are translated to stream only if they are present.
Also, a test for the bug has been added.

Result
------
A request without accept headers is handled properly