THORN-2153: the opentracing-tracerresolver fraction should be internal and stable
Motivation
----------
When running an application that includes the `jaeger` and
`opentracing` fractions, the `opentracing-tracerresolver` fraction
is showed in the list of fractions as unstable.

The fraction should be marked as stable, because all other fractions
that depend on it are also stable, and as internal, because it's not
meant to be used directly.

Modifications
-------------
Marked the `opentracing-tracerresolver` fraction as stable
and internal.

Result
------
Less clutter during application boot.