THORN-2157: make sure we document breaking changes in release announcements
Motivation
----------
The script that fetches the list of fixed issues in given release
could also print the list of all breaking changes in given release
(that is, issues with label `breaking_change`). That of course
won't help with _tracking_ breaking changes, but if we track them,
this should at least help us to not forget about documenting them.

Modifications
-------------
Modified the `fetch-notes.js` script to also include a listing of
all JIRAs that are labeled `breaking_change`.

Added a notice to the release howto to make sure breaking changes
are documented in the release announcement.

Result
------
Extra section in the release announcement with a list of breaking
changes. Hopefully, we won't forget to detail them in writing.