THORN-2175: Jaeger fraction throws RuntimeException for SamplingStrategyResponse
Motivation
----------
When using remote configuration of Jaeger sampling, an exception
is thrown when the Jaeger client is trying to deserialize
the sampling configuration it obtained from the Jaeger server.
This is because the deserializer is using Gson, which is trying
to use `Unsafe` to allocate instance of a class. However, the Gson
module doesn't have a module dependency on `sun.jdk`, so it can't
see the `Unsafe` class.

Modifications
-------------
Add a module dependency to `sun.jdk` to the Gson module.

Result
------
Jaeger client works when remote sampling configuration is used.