THORN-2177: Update MVC-Fraction to v1.0-PFD
Motivation
----------
Some annotations and packages were changed in the latest public final draft of MVC-Spec 1.0.

Modifications
-------------
This updates the MVC-Fraction to the latest version of the spec and it's reference implementation

Result
------
Consumers of the fractions can use the annotations and packages described in the spec. Existing
applications need to update.