[Windows CI Build] Setup different profiles for the TCK test suite.
On Windows, we cannot have case sensitive environment variables. The TCK test suite has mixed case environment variables that cause the CI build (for windows) to fail. This fix creates two profiles - one for windows and one for other environments. In the Windows profile, we do not test for case sensitive environment variables.