THORN-2178: Ability to specify FileSystemLayout via System properties
Motivation
----------
I am trying to fix the Gradle Arquillian adapter. As part of the fixes, I wanted to test out the Gradle project in Thorntail examples. What I noticed was that since the project has a pom.xml as well as a build.gradle file, the FileSystemLayout will always select the Maven layout.

This is a corner case and wouldn't happen in most scenarios, because you are either using Gradle or Maven, very rarely will you use both (unless you are building a product like in the case of the examples repository).

This commit attempts to address this problem by providing the ability to customize the implementation on the basis of a System property.

Modifications
-------------
1. Define an expected signature for FileSystemLayout implementations.
2. Implemented the logic for loading custom file system layouts based on the System property.
3. Implemented a test case that validates different scenarios - valid & invalid.
4. Added minor fixes to support running Arquillian tests in a multi-module Gradle project.

Result
------
The test case proves that the developer can override the layout to use.