THORN-2185: Change config keys to 'thorntail.'
Motivation
----------
It's confusing to have a root config key of 'swarm' when the project is named Thorntail.

Modifications
-------------
Modified config handling to convert 'swarm' to 'thorntail', for backwards compatibility, and updated all config keys to start with 'thorntail'.

Result
------
Config keys now start with 'thorntail', but 'swarm' still works as well.