THORN-2204: non-deterministic module.xml lookup in product with certified fractions
Motivation
----------
Multiple fractions bring the same `module.xml` for the same
module, both in community and product. However, in case of
product with some certified fraction (including `arquillian`),
this is not true -- the product `module.xml` is different from
a community `module.xml`. We are loading the modules by running
`classLoader.getResource("path/to/module.xml")`, which is
non-deterministic in such case: it can randomly load
the productized variant or the community variant (where all
the productized variants available in the classloader are
identical, and also all the community variants are identical).
Choosing the community variant in a product setting can lead
to weird classloading errors (e.g. the difference between
`org.jboss.weld:weld-api` 2.4.SP1 and 2.4.0.SP2-redhat-1).
The product variant (if one exists) should always be loaded,
which is what the `2.2.0.Final-redhat-00021` product release
does. This commit brings the fix back upstream.

Modifications
-------------
When loading `module.xml` from a classloader, always check
if there's a `redhat-` variant, and if so, use it.

Result
------
Loading modules is now deterministic in case of a product with
some certified fractions.