THORN-2262: close the stream of lines on reading Runner cache
Motivation
----------
Thorntail Runner was failing to overwrite previously existing caches
on Windows. It turned out the problem was in the stream of lines
of the cache file not closed after reading.

Modifications
-------------
Stream of lines that is used to read the cache file is now closed after reading.

Result
------
Caches are overwritten properly on Windows

Tested on Windows 10 with IntelliJ IDEA Community Edition