THORN-2264: Runner: pass classpath in env variable
Motivation
----------
On Windows, Thorntail Runner was failing with a project having many dependencies.
The cause was the fact that classpath was passed to FatJarBuilder
with -cp in the command.

Modifications
-------------
The classpath is passed as an environment variable now

Result
------
The Runner no longer fails on a big project
Tested in IntelliJ IDEA and Eclipse on Windows 10