THORN-2277: tracing filter registration no longer overrides other ones
Motivation
----------
OpenTracing, when mixed with JWT, made JWT filters not registered.
The cause was that `TracingInstaller` overwrote RESTEasy providers.
It resulted in an application being unsecured.

Modifications
-------------
Tracing provider is now appended to the already defined providers.
Also, a test that uses the whole microprofile dependency and checks if
MP JWT works was added in a newly created module - testsuite-microprofile

Result
------
MP OpenTracing and JWT can be safely used together now.