THORN-2293: Arquillian tests require the bootstrap module added as a compile time dependency
Motivation
----------
The "findWildFlySwarmBootstrapJar" method in org.wildfly.swarm.tools.DependencyManager searches for the bootstrap jar in the "COMPILE" dependencies only.

When running Arquillian based test cases, it is not necessary that every project has an explicit "compile-time" dependency on the bootstrap module.

E.g.,
In a multi-module project, I might have modules that generate stand-alone libraries, but are still being tested using the Thorntail integration.

Modifications
-------------
Changed the DependencyManager to look for bootstrap jar in test dependencies as well.

Result
------
Defining the Arquillian dependency as a test scope is sufficient to get the test cases running.