THORN-2176: Bootstrap duplicates dependencies
Motivation
----------
JBoss Modules and Snake YAML are compile dependencies to `bootstrap`, but then also shaded into it. This leads to duplicate classes on the classpath and two versions of the classes in the uber jar.

Modifications
-------------
Marked the dependencies as provided scope, so they're only included via shading.

Result
------
Removes duplicate classes on classpath, and JBoss Modules jar is no longer added to m2repo in uber jar.