THORN-1996: realm-name ignored for web.xml login-config
Motivation
----------
If provided in `thorntail.deployment`, the realm-config name for a login-config is always set to "ignored".

Instead we should default it to "ignored" but actually attempt to retrieve a value from YAML.

Modifications
-------------
Modify setting login-config to retrieve a value for `realm-name` key and set to "ignored" if one is not found.

Result
------
Allows a deployment to specify the realm-name for their login-config within YAML.