[THORN-2307] don't generate JavaDocs on Java 9+
The way we generate JavaDocs consists of collecting all the `.java`
files that contain the entire Thorntail public API into a single
directory and then running `javadoc` on that directory.

This used to work fine with Java 8, but fails with Java 11 because
the source code contains references that can't be resolved. In other
words, the extra directory couldn't possibly be compiled, because it
uses unknown symbols. For some reason, that wasn't a problem before
and is now.

One possible solution would be to specify the necessary classpath
that would contain all the necessary classes (the Maven Javadoc
plugin has an option for that), but the list of dependencies is very
long and hard to maintain.

Since the main build stays on Java 8 for the foreseeable future,
building JavaDocs on Java 9+ isn't really required.