[THORN-1594] always open JARs as multi-release
This commit introduces a dependency on the `thorntail-jdk-specific`
artifact. This extra artifact is a multi-release JAR that contains
a few special classes that provide JDK-specific behavior. This JAR
needs to be built with Java 9+, but the rest of Thorntail can stay
on Java 8.

The first JDK-specific behavior that is used is opening JAR files.
When JAR files are opened programmatically, which Thorntail does
very often, a special constructor needs to be used to enable
multi-release behavior. This constructor only exists since Java 9,
so `thorntail-jdk-specific` provides a static method for this.