[THORN-1594] uberjar must be a multi-release JAR
The JBoss Modules JAR and the `thorntail-jdk-specific` JAR are both
multi-release JARs. Since they are directly exploded into the root
of a Thorntail uberjar, the uberjar needs to be a multi-release JAR
as well.

Also the `bootstrap` artifact includes the JBoss Modules JAR, in an
exploded form, so this commit explodes the `thorntail-jdk-specific`
JAR into `bootstrap` and marks `bootstrap` as multi-release.