THORN-2324: add toggle for filtering WEB-INF/lib
Motivation
----------
Provides the ability to deactivate the default filtering of JARs from
WEB-INF/lib when repacking WARs. This avoids situations where the
developer knows that no filtering is required but the plugin removes
required dependencies in error.

Modifications
-------------
A boolean property is added to the build tool and package mojo
controlling whether or not to filter WEB-INF/lib. The property defaults
to true in order to preserve existing behavior by default.

Result
------
Developers wishing to do so can set filterWebinfLib to false in their
poms to disabled fitlering in order to mitigated any issues caused by
the plugin. It will be up to the developer to ensure that any JARs in
WEB-INF/lib belong there.