THORN-2313: Gradle plugin: provide tasks to run|start|stop the application (#1216)
Motivation
----------
The ability to run Thorntail app directly from gradle would (1.) simplify running
functional tests in gradle projects and (2.) improve user experience.

Modifications
-------------
* Tasks "thorntail-[run|start|stop]" were added to the gradle plugin.
* Configuration properties relevant for running the app were added to
  ThorntailExtension of the gradle plugin.

Result
------
It is now possible to run the Thorntail application via "thorntail-run" gradle
task. The "thorntail-start" gradle task can be used to start the application
before integration tests are executed. User should arrange for the
"thorntail-stop" task to be run after integration tests to terminate the
application process. If the stop task is not invoked, the application process
will be terminated at the end of the build.