THORN-2366: documentation still tells users to define "xa-datasource-name" for datasources
Motivation
----------
As we found in the past in THORN-1215, the WildFly's `datasources`
subsystem management model has a usability issue (WFLY-8718) in that
it has two similarly named attributes `xa-datasource-class` and
`driver-xa-datasource-class-name`. The first one is useless, but
our documentation still says on several places that users should
define it. We need to fix the docs to refer to the second attribute
instead.

Modifications
-------------
Replaced all references to `xa-datasource-class`
by `driver-xa-datasource-class-name`.

Result
------
Documentation for defining a JDBC driver is finally correct.