THORN-2411: implement MicroProfile RestClient 1.3
Motivation
----------
MP RestClient 1.3 is part of MP 3.0.

Modifications
-------------
SmallRye RestClient was moved to RESTEasy. So this commit
removes all dependencies on SmallRye RestClient and instead
updates RESTEasy to version 3.8.0 which implements
MP RestClient 1.3. This RESTEasy version is newer than
what WildFly 15.0.1 has, so we need to override some
module definitions.

Result
------
RESTEasy 3.8.0 is used and MP RestClient 1.3 is supported.