THORN-2443: update to RESTEasy 3.8.1
Motivation
----------
This is to align RESTEasy with MP RestClient and future
WildFly release. We use MP RestClient 1.3.3, and RESTEasy 3.8.1
has just updated to this version. We also want to be as close
as possible to which RESTEasy version WildFly 18 will have.
(Can't be 100% sure, but it already has RESTEasy 3.8.1.)

Modifications
-------------
Update to RESTEasy 3.8.1.

Result
------
Better RESTEasy alignment with MP RestClient and WildFly.
No behavioral changes.