THORN-2458: align Apache HTTP Client versions with WildFly
Motivation
----------
The versions of Apache HTTP Client we use are out of sync
with WildFly, which leads to potentially multiple versions
of it being present in an uberjar. We should align
with WildFly, resp. WildFly Core. And since Thorntail is
currently still based on WildFly 15.0.1.Final, we need to
align with WildFly Core 7.0.0.Final [1].

[1] https://github.com/wildfly/wildfly-core/blob/7.0.0.Final/pom.xml#L155-L156

Modifications
-------------
Use the same versions of Apache HTTP Client as WildFly uses.

Also add dependency management of several other libraries
that also need to be aligned with WildFly.

Result
------
No behavioral change. Slightly easier maintenance.