THORN-2460: thorntail.classpath system property can be used also for uberjars
Motivation
----------
Documentation is not very clear about using external JDBC drivers.
It suggests that the thorntail.classpath property can be used only
with hollow JARs, which is not true. Improve wording so it is clear
that it can be used for both hollow JARs and uberjars.

Modifications
-------------
Reworded and renamed the section about specifying JDBC drivers
via the thorntail.classpath system property.

Result
------
Better user experience when reading the documentation.