THORN-2466: make the JsonMapping test pass
The test as was originally written didn't pass, because the custom
Jackson provider was declared like this:

```java
@Consumes(MediaType.WILDCARD)
@Produces(MediaType.WILDCARD)
```

At the same time, the `jaxrs` fraction automatically brings in
the RESTEasy's builtin Jackson provider, which is declared like

```java
@Consumes({"application/json", "application/*+json", "text/json"})
@Produces({"application/json", "application/*+json", "text/json"})
```

Per the JAX-RS sorting rules, wildcard providers come _after_
providers with concrete media types, even if the wildcard provider
is custom and the specific one is built in. (At least that's
the current RESTEasy behavior, and I trust it's conformant
with the spec. I'm no JAX-RS expert, so can't personally tell.)

The solution is easy: let the custom provider declare the same
`@Consumes` and `@Produces` annotations as the builtin one.