Gradle: Align the Gradle version during Arquillian dependency resolution
Previously, the Gradle Arquillian integration would always spin up a Gradle 3.1 daemon. This was becuase we were pulling in the Gradle jar (version 3.1) as a dependency. This has a performance implication on the CI equation as each build would download Gradle 3.1 for testing.

With this commit, I am making the following changes,

1. Mark all Gradle dependencies as provided.
2. Include the required dependencies (from the Gradle platform) dynamically as part of the plugin configurations.
3. Introduce a path forward for adding Gradle specific changes without having to rely on bumping the Gradle library versions.