Avoid replacing the original war file if no filtering is required
The current implementation will always replace the built war file irrespective of whether the `filterWebInfLib` flag has been set or not. By doing an early check we can speed up the execution time for projects that build large war files.

Additionally, this would also help the Gradle plugin leverage task avoidance which is currently not happening. By replacing the war file (with the same content), Gradle believes that the file has changed and thus re-executes the war task, which will re-execute the thorntail-package task. With this change, the Gradle plugin will be able to avoid repackaging and speed up development.