THORN-2492: Maven plugin ignores exclusions when creating wildfly-swarm-manifest.yaml
Motivation
----------
When the `pom.xml` declares an exclusion for a library, this exclusion
is not taken into account when the `wildfly-swarm-manifest.yaml` file
is created. This can cause library conflicts at runtime.

Example:
The SLF4J logging framework supports different underlying logging systems,
e.g. Log4j or Logback; however, only one of the corresponding binding JAR
files is allowed to be present at runtime.

When a Thorntail project uses the SLF4J logging framework, declares
a direct dependency on one of the bindings, and declares a dependency
on a 3rd party library (e.g. ZooKeeper) which itself declares a dependency
on another binding, this will lead to multiple bindings being present
at runtime. The 3rd party library should normally not declare that dependency,
but if it does, the only solution is to add an exclusion; however, this is
ignored when the `wildfly-swarm-manifest.yaml` file is created, so the SLF4J
system complains at runtime about duplicate bindings.

Modifications
-------------
When the uberjar is being built by the Maven plugin, we can access
the entire dependency tree exactly how Maven sees it. Therefore, we don't
have to figure out the transitive dependencies ourselves, which has a nice
benefit: dependency exclusions are taken into account.

Result
------
Only affects how the Maven plugin builds the uberjar. Other code paths
that build an uberjar are not affected, which unfortunately means
there might be behavioral differences between e.g. Maven plugin
and Arquillian.