THORN-2508: jaxrs-jsonb fraction doesn't bring in jaxrs-jsonp
Motivation
----------
The `jsonb` fraction brings in the `jsonp` fraction transitively.
It isn't the case for the `jaxrs-jsonb` fraction -- it currently
doesn't bring in `jaxrs-jsonp`. I believe it should, because when
working with JSON-B, you very often also need JSON-P.

Modifications
-------------
Added a dependency on `jaxrs-jsonp` to the `jaxrs-jsonb` fraction.

Result
------
Made the work with JSON-B in JAX-RS a little easier.