Gradle Plugin: Optimize lookup for module dependencies in DependencyManager
The ModuleAnalyzer class looks for dependencies in a predefined location and assumes that the folder conforms
to the Maven repository layout. If it doesn't find the dependency in that location, then the ModuleAnalyzer will
return an ArtifactSpec object that does not contain the file location. But when using the Gradle plugin, it
is highly possible that the dependency was already retrieved and stored in the Gradle dependency cache.

In this commit, I am performing an additional check to handle such scenarios.