THORN-2526: add option to only repackage WAR inside the uberjar and leave the WAR file untouched
Motivation
----------
When `<packaging>` is `war`, we filter the `WEB-INF/lib` directory
of the WAR file by default. We do it on two places -- in the WAR
which is inside the uberjar, and in the standalone WAR file.

This is inconvenient when we want to build the uberjar, but also keep
using the WAR file e.g. for deploying to WildFly. In such scenario,
we want to modify the WAR which is inside the uberjar, but keep
the standalone WAR file untouched.

Modifications
-------------
This could be achieved by adding another option, e.g. `repackageWar`,
but the combination of `repackageWar` and `filterWebInfLib` would be
confusing. So this commit adds a 3rd value to `filterWebInfLib`:
`uberjar-only`. With this value, the standalone WAR file is left
untouched, only the WAR inside the uberjar is modified.

Result
------
No behavioral difference by default.
Added the ability to only modify the WAR inside the uberjar,
but not modify the standalone WAR file.