[THORN-2537] Upgrade dependencies of MVC fraction
Motivation
---------
This commit upgrades the dependencies to the MVC API and reference
implementation to replace the outdated versions with the latest released
ones.

Modifications
-------------
This commit modifies just build data (pom.xml of parent project and
fractions) as well as metadata of the module definitions. There is no
change in productive java code.

Result
------
With this commit, people using Thorntail can use the latest and stable
versions of the MVC API and its RI.