[THORN-2537] Mark MVC fraction as stable
Motivation
----------
To signal people using the fraction that it can be used without expecting bigger API
changes, the stability should be set to 'stable'.

Modifications
-------------
This commit changes only the 'swarm.fraction.stability' value in the pom.xml to 'stable'.

Result
------
The fraction is displayed as stable to people using it.