THORN-2538: Use of thorntail.runner.webapp-location generates a invalid WAR
Motivation
----------
When pointing the `thorntail.runner.webapp-location` system property
to a _target_ directory (such as `target/my-app-name`), as opposed to
a _source_ directory (such as `src/main/webapp`), the generated WAR
is invalid WAR.

This is useful when some post-compilation tasks are done by Maven
(such as Maven overlays or custom Ant tasks) and should be used
when generating the Thorntail runner WAR.

Modifications
-------------
The `addWebappResourceToWar` method used to ignore the configured
webapp path and always used `src/main/webapp`. Now, it uses
the configured path correctly.

Also, the method now filters out `WEB-INF/classes` and `WEB-INF/lib`,
because these are not webapp resources. The `.class` and `.jar` files
are added to the WAR elsewhere.

Result
------
Possible to set `thorntail.runner.webapp-location` to a _target_ path.