dependency updates
- update to RESTEasy 3.9.3.SP1
- update to Jackson 2.10.2
- update to Jaeger client 0.34.1 and Apache Thrift 0.13.0
- update to OpenShift Rest Client Java 9.0.0.Final
- update to SnakeYAML 1.24

Jackson 2.10 needs up to date SnakeYAML (1.24). This version adds
some calls to `java.util.logging`. At the same time, Thorntail
bootstrap uses SnakeYAML early and so it is required for SnakeYAML
to not touch `java.util.logging`, which would prevent `logging`
subsystem startup.

The fix is to update SnakeYAML, but fork some of its classes
and remove the `java.util.logging` calls (fortunately the number
of classes that touch `java.util.logging` is very low), just like
what we do with JBoss Modules. We then make sure that everyone
uses SnakeYAML bundled in the Thorntail bootstrap. This has always
been the case when running uberjar, because SnakeYAML is consumed
via module dependency, and the SnakeYAML module we provide actually
takes SnakeYAML from the bootstrap. When running from classpath,
we just need to make sure that bootstrap is on classpath sooner
than SnakeYAML, which is easy to do in the Maven plugin.